requirements.txt
libigl==2.4.1
numpy~=1.26.0
potpourri3d==1.0.0
PyYAML==6.0.1
robust_laplacian==0.2.7
scikit_learn==1.3.1
scipy~=1.11.3
torch==2.0.0
torch_geometric==2.3.1
tqdm==4.65.0
tensorboard
scikit-learn~=1.3.1
pandas~=2.2.0
vtk~=9.3.0
pymeshfix~=0.16.3